pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tableshot"
version = "0.1.0"
description = "Extract tables from PDFs into clean, structured data -- instantly. An MCP server for AI assistants."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Andrew Makris" }]
keywords = ["mcp", "table-extraction", "pdf", "document-ai", "structured-data", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: General",
]

dependencies = [
    "mcp>=1.0",
    "pdfplumber>=0.10",
    "pypdfium2>=4.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
ml = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.30",
    "timm>=0.9",
]
ocr = [
    "onnxtr[cpu]>=0.5",
]
all = [
    "tableshot[ml,ocr]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "fpdf2>=2.7",
]

[project.urls]
Homepage = "https://github.com/Bespoke34/tableshot"
Repository = "https://github.com/Bespoke34/tableshot"
Issues = "https://github.com/Bespoke34/tableshot/issues"

[project.scripts]
tableshot = "tableshot.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tableshot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100
